requirements.txt
# Database
psycopg2-binary==2.9.9

# AI / LLM
anthropic==0.39.0

# Speech-to-Text (Whisper runs locally)
openai-whisper==20231117
torch>=2.0.0

# Audio capture
sounddevice==0.4.6
soundfile==0.12.1
numpy>=1.24.0

# Utilities
python-dotenv==1.0.0
rich==13.7.0          # Pretty console output
tabulate==0.9.0       # Table formatting

# Development
pytest==7.4.0
